+++ src/utils/getIsSortEnabled.ts
@@ -1,0 +1,5 @@
+import { CLIOptions, options } from "../cli/options";
+import config from "../config";
+
+export const getIsSortEnabled = (): boolean =>
+  config.sortChangelog || options[CLIOptions.sortOnly];