+++ changelog.md
@@ -1,3 +1,5 @@
+# [vNext]
+
 # [v1.0.4] 2023-02-07
 
 - [PATCH] Fix CRLF issue on Mac/Linux